Refactor into multiple components
